# docs/spec/04_M1_OpenAPI.yaml
openapi: 3.1.0
info:
  title: 같이놀개 M1 API
  version: 1.3.1-m1
  description: |
    최신 제품 정책과 M1 통합 ERD를 기준으로 작성한 정적 REST 계약 정본이다.
    런타임 /v3/api-docs는 현재 구현 관찰용이며 계약 정본이나 필수 CI Gate가 아니다.
    M1은 폴링 방식이며 WebSocket, Push, 읽음 표시, 메시지 수정·삭제,
    사용자 셋로그 업로드, 회원탈퇴, 욕설 자동 차단 및 AI 대화 검열을 제공하지 않는다.
    최상위 bearerAuth가 적용되는 모든 operation은 인증 실패 시
    401 ErrorEnvelope를 반환한다.
    별도 HTTP Status는 이 정적 계약을 먼저 변경하고 ErrorCode·구현·테스트를 같은
    PR에서 동기화한다. 소유권 실패는 기본 403이며 존재 은닉이 필요한 endpoint만
    명시적으로 404를 사용한다.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /pets:
    post:
      tags: [Pet]
      summary: Pet 직접 등록
      operationId: createPet
      description: |
        owner User를 직렬화한 생성 트랜잭션에서 deleted_at IS NULL Pet 수를 조회한다.
        이 생성 트랜잭션에서 owner User 잠금·동시 수정 충돌이 발생하면 Pet을
        Commit하지 않고 409 CONCURRENT_UPDATE_CONFLICT를 반환한다.
        생성 전 수가 0이면 내부 firstPetCandidate=true로 확정하고 Pet을 Commit한다.
        생성 Commit 뒤 후보인 경우에만 별도 트랜잭션으로 Active 지정을 시도한다.
        자동 Active 지정 단계에서 비관적 잠금 실패로 분류된 동시성 오류에는 생성된
        Pet을 유지하고 201과 RETRY_REQUIRED를 반환한다. 자동 지정의 DB 연결 장애·
        무결성 오류·코딩 오류·불변조건 위반은 RETRY_REQUIRED 또는 NOT_APPLICABLE로
        숨기지 않고 원인에 맞는 오류 흐름으로 처리한다. 선행 Pet 생성이 이미
        Commit됐다면 후속 자동 Active 지정이 오류로 실패해도 생성 Pet은 유지된다.
        같은 owner의 미삭제 Pet은 SUSPENDED를 포함해 최대 5마리이며 초과 시
        409 PET_LIMIT_EXCEEDED를 반환한다.
        Pet PublicTag Unique 충돌로 총 5회 저장에 실패하면
        409 PET_PUBLIC_TAG_GENERATION_FAILED를 반환한다.
        M1은 Idempotency-Key를 제공하지 않는다. timeout·5xx 뒤에는 GET /pets/me로
        생성 여부를 확인하고 POST를 무작정 재시도하지 않는다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetCreateRequest'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '201':
          description: Pet 등록 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetCreateEnvelope'
        '409':
          description: PET_LIMIT_EXCEEDED, PET_PUBLIC_TAG_GENERATION_FAILED 또는 CONCURRENT_UPDATE_CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /pets/{petId}:
    parameters:
      - $ref: '#/components/parameters/PetId'
    get:
      tags: [Pet]
      summary: Pet 상세 조회
      operationId: getMyPet
      description: |
        본인 소유의 미삭제 Pet 전체 정보를 조회한다.
        Pet 행 존재 여부, 삭제 여부, 미삭제 Pet의 소유권 순서로 확인한다.
        Pet 행이 없거나 삭제된 Pet이면 소유권 검사보다 먼저
        404 PET_NOT_FOUND를 반환한다.
        삭제되지 않은 Pet이 존재하지만 다른 User 소유인 경우에만
        403 PET_NOT_OWNED를 반환한다.
        따라서 다른 User 소유이면서 삭제된 Pet도 404 PET_NOT_FOUND로 처리한다.
        과거 Chat 이력 표시용 내부 Pet 조회는 이 REST endpoint와 별도이며
        삭제된 Pet도 조회할 수 있다.
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Pet 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetEnvelope'
        '404':
          description: Pet 행 없음 또는 삭제. 소유권 검사보다 우선한다. (PET_NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: 삭제되지 않은 다른 User 소유 Pet (PET_NOT_OWNED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [Pet]
      summary: Pet 수정
      operationId: updatePet
      description: |
        인증 User가 소유한 미삭제 Pet 정보를 부분 수정한다. 현재 Active Pet이
        아니거나 SUSPENDED 상태여도 허용한다. 생략 필드는 유지하고 nullable 필드의
        명시적 null은 값을 초기화한다. nickname과 personalityTags의 null은 허용하지
        않으며 personalityTags의 빈 배열은 전체 태그를 제거한다.
        HTTP Body 없음, JSON null, 빈 객체, unknown field와 수정 불가 필드는
        VALIDATION_FAILED다. PATCH는 필드별 JSON wire 타입을 엄격하게 검사하며
        문자열·숫자·boolean 간 coercion을 하지 않는다. nullable 문자열의 textual
        value는 생성 계약과 같이 trim하거나 빈 값을 null로 바꾸지 않는다.
        Pet 행 없음 또는 삭제를 소유권보다 먼저 PET_NOT_FOUND로 처리하고, 미삭제
        타인 소유 Pet은 PET_NOT_OWNED로 처리한다. 실제 값이 같은 no-op도 200이며
        version 또는 updatedAt 갱신을 강제하지 않는다.
        현재 Verification 저장 구조가 없어 응답은 verified=false,
        verifiedAt=null이다. 향후 Verification 구현 뒤 nickname, breedName, sex,
        neutered, birthDate가 인증 스냅샷과 달라지면 같은 트랜잭션에서 배지를
        해제해야 한다. 이번 API는 신규 Flyway 또는 ERD 변경을 요구하지 않는다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PetUpdateRequest'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
        '200':
          description: 수정 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetEnvelope'
        '403':
          description: 미삭제 타인 소유 Pet (PET_NOT_OWNED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Pet 행 없음 또는 삭제. 소유권 검사보다 우선한다. (PET_NOT_FOUND)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: 낙관적 잠금 충돌 (CONCURRENT_UPDATE_CONFLICT)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [Pet]
      summary: Pet 삭제
      operationId: deletePet
      description: Active Pet은 삭제할 수 없다.
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '204':
          description: 삭제 완료
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /pets/{petId}/friends:
    get:
      tags: [Friend]
      summary: Pet 친구 목록
      operationId: listPetFriends
      description: |
        인증 User가 소유한 미삭제 Pet의 친구를 최신 Friendship 순서로 반환한다.
        현재 Active Pet이 아니거나 SUSPENDED 상태여도 조회할 수 있다.
        다른 User 소유 Pet은 PET_NOT_OWNED, 없거나 삭제된 Pet은 PET_NOT_FOUND다.
      parameters:
        - $ref: '#/components/parameters/PetId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 친구 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PetSearchListEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /pets/{petId}/friends/{friendPetId}:
    delete:
      tags: [Friend]
      summary: 친구 삭제
      operationId: deletePetFriend
      description: |
        인증 User가 소유한 미삭제 source Pet의 Friendship만 삭제한다.
        source Pet은 현재 Active Pet일 필요가 없으며 ACTIVE 또는 SUSPENDED를 허용한다.
        source Pet이 없거나 삭제됐으면 PET_NOT_FOUND, 미삭제 타인 소유 Pet이면
        PET_NOT_OWNED, canonical pair의 Friendship이 없으면 FRIENDSHIP_NOT_FOUND다.
        target Pet은 선행 조회하지 않는다. 기존 DIRECT ChatRoom, Participant,
        Message, Greeting, FriendRequest 이력은 변경하지 않으며 기존 Chat 조건을
        충족하는 참가자는 계속 대화할 수 있다.
      parameters:
        - $ref: '#/components/parameters/PetId'
        - name: friendPetId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '204':
          description: 친구 삭제 완료. 응답 Body 없음
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 보호 API의 인증 실패는 401 ErrorEnvelope로 반환한다.
  parameters:
    PetId:
      name: petId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: 권한 또는 정책 위반
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: 상태 또는 중복 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
        data:
          type: 'null'
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    PetCreateRequest:
      type: object
      required: [nickname]
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 30
          description: >-
            앞뒤 공백 제거 후 저장하며, 이모지 및 emoji-like pictographic
            문자·기호(예: ©, ™, ☀, ♥)는 사용할 수 없다.
        breedName:
          type: [string, 'null']
          maxLength: 100
          description: 사용자 입력 또는 향후 동물등록 조회의 kindNm을 반영할 수 있는 견종명
        sex:
          $ref: '#/components/schemas/NullablePetSex'
        neutered:
          type: [boolean, 'null']
        birthDate:
          type: [string, 'null']
          format: date
        weightKg:
          type: [number, 'null']
          minimum: 0
          maximum: 999.99
          multipleOf: 0.01
        sizeCode:
          type: [string, 'null']
          enum: [SMALL, MEDIUM, LARGE, null]
        bio:
          type: [string, 'null']
          maxLength: 500
        personalityTags:
          type: array
          maxItems: 10
          items:
            type: string
        careNote:
          type: [string, 'null']
          maxLength: 500
    PetUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      description: |
        필드 생략은 기존 값 유지, nullable 필드의 null은 초기화를 의미한다.
        nickname과 personalityTags는 null을 허용하지 않으며 personalityTags의
        빈 배열은 전체 제거다. minProperties는 정적 표현이며 런타임 parser가
        지원 필드 presence와 unknown field를 다시 검사한다.
        생성 API의 값 정규화 정책을 유지하지만 PATCH JSON field는 schema의 실제
        wire 타입만 허용하며 scalar coercion을 하지 않는다.
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 30
          description: >-
            앞뒤 공백 제거 후 저장하며, 이모지 및 emoji-like pictographic
            문자·기호(예: ©, ™, ☀, ♥)는 사용할 수 없다. null은 허용하지 않는다.
        breedName:
          type: [string, 'null']
          maxLength: 100
          description: >-
            사용자 입력 또는 향후 동물등록 조회의 kindNm을 반영할 수 있는 견종명.
            textual value는 trim하지 않고 빈 문자열과 공백 문자열도 그대로 저장한다.
        sex:
          $ref: '#/components/schemas/NullablePetSex'
        neutered:
          type: [boolean, 'null']
        birthDate:
          type: [string, 'null']
          format: date
        weightKg:
          type: [number, 'null']
          minimum: 0
          maximum: 999.99
          multipleOf: 0.01
        sizeCode:
          type: [string, 'null']
          enum: [SMALL, MEDIUM, LARGE, null]
        bio:
          type: [string, 'null']
          maxLength: 500
          description: textual value는 trim하거나 빈 문자열을 null로 바꾸지 않는다.
        personalityTags:
          type: array
          maxItems: 10
          description: >-
            null은 허용하지 않는다. 빈 배열은 전체 제거이며 요소를 trim, 정렬,
            중복 제거하지 않는다.
          items:
            type: string
        careNote:
          type: [string, 'null']
          maxLength: 500
          description: textual value는 trim하거나 빈 문자열을 null로 바꾸지 않는다.
    NullablePetSex:
      type: [string, 'null']
      enum: [MALE, FEMALE, UNKNOWN, null]
    Pet:
      type: object
      required: [petId, ownerUserId, publicTag, nickname, status, deletedAt, verified, active]
      properties:
        petId:
          type: integer
          format: int64
        ownerUserId:
          type: integer
          format: int64
        publicTag:
          type: string
          maxLength: 30
          pattern: '^.{1,25}#[A-Z0-9]{4}$'
          description: User와 별도 Namespace. trim한 nickname 앞 25개 Unicode code point와 #XXXX로 생성한다.
        ownerPublicTag:
          type: string
        nickname:
          type: string
          minLength: 1
        breedName:
          type: [string, 'null']
          description: 사용자 입력 또는 향후 동물등록 조회의 kindNm을 반영할 수 있는 견종명
        sex:
          $ref: '#/components/schemas/NullablePetSex'
        neutered:
          type: [boolean, 'null']
        birthDate:
          type: [string, 'null']
          format: date
        weightKg:
          type: [number, 'null']
        sizeCode:
          type: [string, 'null']
          enum: [SMALL, MEDIUM, LARGE, null]
        bio:
          type: [string, 'null']
        personalityTags:
          type: array
          items:
            type: string
        careNote:
          type: [string, 'null']
        profileUrl:
          type: [string, 'null']
          format: uri
          description: M1 사용자 등록 Pet은 null이며 클라이언트가 기본 이미지를 표시한다.
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, DELETED]
          description: DELETED이면 deletedAt이 필수이며 ACTIVE/SUSPENDED이면 deletedAt은 null이다.
        deletedAt:
          type: [string, 'null']
          format: date-time
        verified:
          type: boolean
        verifiedAt:
          type: [string, 'null']
          format: date-time
        active:
          type: boolean
    PetSearchItem:
      type: object
      required: [petId, publicTag, nickname, verified, relationship]
      properties:
        petId:
          type: integer
          format: int64
        publicTag:
          type: string
          maxLength: 30
          pattern: '^.{1,25}#[A-Z0-9]{4}$'
          description: Pet 공개 태그
        nickname:
          type: string
          minLength: 1
        profileUrl:
          type: [string, 'null']
          format: uri
          description: M1은 null이며 클라이언트가 기본 이미지를 표시한다.
        verified:
          type: boolean
        relationship:
          type: [string, 'null']
          enum: [NONE, REQUEST_SENT, REQUEST_RECEIVED, FRIEND, null]
          description: Active Pet이 없는 L1은 null
...
    CursorPage:
      type: object
      required: [nextCursor, hasNext]
      properties:
        nextCursor:
          type: [string, 'null']
        hasNext:
          type: boolean
...
    PetEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: '#/components/schemas/Pet' }
        error: { type: 'null' }
    PetCreateResult:
      type: object
      required: [pet, activeAssignmentStatus]
      properties:
        pet:
          $ref: '#/components/schemas/Pet'
        activeAssignmentStatus:
          type: string
          enum: [ASSIGNED, RETRY_REQUIRED, NOT_APPLICABLE]
          description: |
            firstPetCandidate=true이고 자동 지정에 성공하면 ASSIGNED,
            후보지만 자동 지정 단계에서 비관적 잠금 실패로 분류된 동시성 오류면
            RETRY_REQUIRED,
            firstPetCandidate=false이거나 이미 Active Pet이 있으면 NOT_APPLICABLE이다.
            DB 장애·무결성 오류·코딩 오류는 RETRY_REQUIRED로 숨기지 않는다.
            firstPetCandidate 자체는 외부 응답에 노출하지 않는다.
    PetCreateEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: '#/components/schemas/PetCreateResult' }
        error: { type: 'null' }
...
    PetSearchListEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data:
          type: object
          required: [items, page]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/PetSearchItem' }
            page: { $ref: '#/components/schemas/CursorPage' }
        error: { type: 'null' }
...